LayoutTests/platform/wpe/imported/w3c/web-platform-tests/fetch/metadata/style.https.sub-expected.txt

Blocked access to external URL https://www.localhost:9443/fetch/metadata/resources/record-header.py?file=style-same-site7d707274-c421-4635-ab65-1b1670c512b0

Harness Error (TIMEOUT), message = null

PASS Same-Origin style
TIMEOUT Same-Site style Test timed out
NOTRUN Cross-Site style
NOTRUN Same-Origin, cors style
PASS Same-Origin style: sec-fetch-dest
PASS Same-Origin style: sec-fetch-mode
FAIL Same-Origin style: sec-fetch-site assert_equals: expected "same-origin" but got ""
PASS Same-Origin style: sec-fetch-user
